First try: Streaming from subprocess
This works... sort of. The template is updated as the content comes from
the process, but not all process output is gulped by the iterator. The
output from latex is not catched, for instance.

There might also be security problems related to this (as the subprocess
uses shell=True). That might be solvable, if it even is a problem, but
it should be looked into.

Finally, it is very unsatisfying to drop into a subprocess and construct
a faux CLI call when you have all the utilities available from the
module import. This feels like a very unpythonic solution (aside from
being only halfway good).