Implementation of web sockets in stead of yielding stream
This gives fast asynchronous server-client communication. This solution
sidesteps the request procedure generally used, meaning that updates
from the server can be streamed directly to the client without any
client requests or template reloads. The major downside is that it
requires javascript enabled to work. This solution is totally dependent
on javascript, and it might be a problem. I don't know if we want to
move away from it eventually, but I think the users and our own use and
development of it will show.