Add local version of reledmac and reledpar
It is necessary as texlive got frozen just before publication of
important update.

See https://github.com/maieul/ledmac/issues/634

But this also means that updates of reledmac must be done to these local
files rather than through tlgmgr. This should only be temporary, viz.
until we move to texlive 2017, but that might be a while.